requirements.txt
python-jose
passlib
httpx
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
